adds pagination and refactors existing methodology pagination buttons
